+++ R/weight2rows.R
@@ -1,16 +1,50 @@
 #' Expand a weighted data frame to an equivalent unweighted
 #' @param DT A \code{data.table}. Will be converted to one if possible.
 #' @param weight.var Variable in \code{DT} to be used as weights.
+#' @param rows.out If not \code{NULL} (the default) specifies the number of rows in the result;
+#' otherwise the number of rows will be \code{sum(DT[[weight.var]])}. (Due to rounding, this figures are inexact.)
 #' @return \code{DT} but with the number of rows expanded to \code{sum(DT[[weight.var]])} to reflect the weighting.
 #' @examples 
 #' 
 #' library(data.table)
 #' DT <- data.table(x = 1:5, y = c(1, 1, 1, 1, 2))
 #' weight2rows(DT, "y")
+#' weight2rows(DT, "y", rows.out = 5)
 #' 
 #' @export 
 
-weight2rows <- function(DT, weight.var) {
+weight2rows <- function(DT, weight.var, rows.out = NULL) {
+  if (!is.data.table(DT)) {
+    if (!is.data.frame(DT)) {
+      stop("`DT` was a ", class(DT)[1L], ". ",
+           "DT must be a data.frame.")
+    }
+    DT <- as.data.table(DT)
+  }
+  
+  the_colorder <- copy(names(DT))
+  
+  
+  if (length(weight.var) != 1L) {
+    stop("`weight.var` had length ", length(weight.var), ". ",
+         "`weight.var` must be a single column name or position.")
+  }
+  if (is.numeric(weight.var)) {
+    if (weight.var < 1 || weight.var > ncol(DT)) {
+      stop("`weight.var = ", weight.var, "` outside the ", 
+           "range 1 <= x <= ", ncol(DT), " = ncol(DT).\n",
+           "If used to specify a position, `weight.var` must specify a column number.")
+    }
+    weight.var <- names(DT)[weight.var]
+  } else if (is.character(weight.var)) {
+    if (weight.var %notchin% names(DT)) {
+      stop("`weight.var = ", weight.var)
+    }
+  } else {
+   stop("`typeof(weight.var) = '", typeof(weight.var), "'`, ",
+        "but must be numeric or character.") 
+  }
+  
   weight.var.value <- DT[[weight.var]]
   if (anyNA(weight.var.value)) {
     warning("`weight.var` contained NAs. These have been converted to zeroes.")
@@ -24,25 +58,62 @@ weight2rows <- function(DT, weight.var) {
          "Modify `weight.var` so that all the values are nonnegative.")
   }
   
-  if (!is.data.table(DT)) {
-    setDT(DT)
+  
+  
+  if (is.null(rows.out)) {
+    M <- 1L
+  } else {
+    if (!is.numeric(rows.out)) {
+      stop("`typeof(rows.out) = '", typeof(rows.out), "'`.\n",
+           "`rows.out`, if used, must be a single number.")
+    }
+    if (length(rows.out) != 1L) {
+      stop("`length(rows.out) = ", length(rows.out), "`.\n",
+           "`rows.out`, if used, must be a single number.")
+    }
+    if (anyNA(rows.out)) {
+      stop("`rows.out = NA` but NA is not permitted.\n", 
+           "`rows.out`, if used, must be a single number.")
+    }
+    M <- rows.out / sum(weight.var.value)
   }
   
-  switch(typeof(weight.var.value), 
-         "logical" = {
-           warning("weight.var is logical. Treating as filter/subset.")
-           DT[which(weight.var.value)]
-         },
-         "integer" = {
-           DT %>%
-             .[weight.var.value > 0] %>%
-             .[, lapply(.SD, rep_len, .BY[[1]] * .N), .SDcols = names(.)[names(.) != weight.var], by = weight.var]
-         },
-         "double" = {
-           DT %>%
-             .[weight.var.value > 0] %>%
-             .[, lapply(.SD, rep_len, round(.BY[[1]] * .N)), .SDcols = names(.)[names(.) != weight.var], by = weight.var]
-         }, 
-         stop("Non-numeric weight.var. Aborting."))
-
+  rep_out <- function(x, BY, N, M) {
+    len <- BY * N * M
+    if (len < 1) {
+      len <- if (runif(1) < len) 1L else 0L
+    } else {
+      len <- as.integer(len)
+    }
+    rep_len(x, len)
+  }
+  
+  
+  
+  out <- 
+    switch(typeof(weight.var.value), 
+           "logical" = {
+             warning("weight.var is logical. Treating as filter/subset.")
+             DT[which(weight.var.value)]
+           },
+           "integer" = {
+             DT %>%
+               .[weight.var.value > 0] %>%
+               .[, lapply(.SD, rep_out, .BY[[1]], .N, M),
+                 .SDcols = names(.)[names(.) != weight.var],
+                 by = weight.var]
+           },
+           "double" = {
+             DT %>%
+               .[weight.var.value > 0] %>%
+               .[, lapply(.SD, rep_out, .BY[[1]], .N, M),
+                 .SDcols = names(.)[names(.) != weight.var],
+                 by = weight.var]
+           }, 
+           stop("Non-numeric weight.var. Aborting."))
+  
+  # by will fix things first
+  setcolorder(out, the_colorder)
+  out[]
+  
 }